Added --include-path and --library-path to setup options
